data_layer/requirements.txt
flask
pyjwt
datetime
uuid
Flask-SQLAlchemy
flask-expects-json
